api/requirements.txt
Django>=3.0,<4.0 # Django framework
psycopg2>=2.8 # for postgreSQL
djangorestframework==3.13.1 # Django Rest Framework